refactor(openapi): no defend missing key in openapi and separate opening file to the class FileOpener (#5)
