+++ luke/openapi.py
@@ -2,9 +2,9 @@ import logging
 import json
 from typing import List, Dict, Any, Tuple, Optional
 import yaml
-import httpx
 import jsonschema
 from .schemas import openapi3_0_schema
+from .file_opener import FileOpener
 
 
 logger = logging.getLogger("luke")
@@ -33,24 +33,8 @@ class OpenAPISpec:
                     logger.warn(f"Ignore path {method}:{path} because of {str(e)}")
 
     @classmethod
-    def open_from_url(cls, url: str) -> str:
-        with httpx.Client() as client:
-            response = client.get(url, timeout=10)
-            response.raise_for_status()    
-            return response.text
-
-    @classmethod
-    def open_from_file(cls, filename: str):
-        with open(filename, "r") as f:
-            return f.read()
-
-    @classmethod
     def load_file(cls, filename_or_url: str) -> dict:
-        content = None
-        if filename_or_url.startswith("https://") or filename_or_url.startswith("http://"):
-            content = cls.open_from_url(filename_or_url)
-        else:
-            content = cls.open_from_file(filename_or_url)
+        content = FileOpener.open_from_url(filename_or_url)
 
         if not content:
             raise ValueError("File is empty")
@@ -64,20 +48,24 @@ class OpenAPISpec:
 
     @classmethod
     def validate_openapi(cls, spec: dict):
-        return jsonschema.validate(spec, openapi3_0_schema) 
+        return jsonschema.validate(spec, openapi3_0_schema)
 
     def resolve_spec(self, spec: dict) -> dict:
         resolved = spec
         ref = spec.get("$ref")
         if ref:
             resolved = self.resolve_ref(ref)
-        
+
         if resolved["type"] == "object":
             if "properties" in resolved:
                 for key in resolved["properties"]:
-                    resolved["properties"][key] = self.resolve_spec(resolved["properties"][key])
+                    resolved["properties"][key] = self.resolve_spec(
+                        resolved["properties"][key]
+                    )
             else:
-                resolved["additionalProperties"] = self.resolve_spec(resolved["additionalProperties"])
+                resolved["additionalProperties"] = self.resolve_spec(
+                    resolved["additionalProperties"]
+                )
 
         elif resolved["type"] == "array":
             resolved["items"] = self.resolve_spec(resolved["items"])
@@ -97,10 +85,9 @@ class OpenAPISpec:
 
         return node
 
-
     @property
     def info(self) -> dict:
-        return self.spec.get("info", {})
+        return self.spec["info"]
 
 
 class Endpoint:
@@ -126,14 +113,14 @@ class Endpoint:
         content_specs = self.content_specs
         for code, contents in spec["responses"].items():
             if "content" not in contents:
-                content_specs[(code, "application/json")] = {
-                    "type": "string"
-                }
+                content_specs[(code, "application/json")] = {"type": "string"}
                 continue
 
             for content_type, content in contents["content"].items():
                 try:
-                    content_specs[(code, content_type)] = self.openapi.resolve_spec(content["schema"])
+                    content_specs[(code, content_type)] = self.openapi.resolve_spec(
+                        content["schema"]
+                    )
                 except KeyError:
                     pass
 
@@ -141,7 +128,7 @@ class Endpoint:
         if "headers" in spec:
             for header_name, header_spec in spec["headers"].items():
                 try:
-                    headers_spec["properties"][header_name] = self.openapi.resolve_spec(header_spec["schema"]) # type: ignore
+                    headers_spec["properties"][header_name] = self.openapi.resolve_spec(header_spec["schema"])  # type: ignore
                 except KeyError:
                     continue
